+++ src/81commit.js
@@ -5,12 +5,12 @@
 // (c) 2014, Andrey Gershun
 //
 */
-yy.Begin = function (params) { return yy.extend(this, params); }
-yy.Begin.prototype.toString = function() {
-	return K('BEGIN');
+yy.BeginTransaction = function (params) { return yy.extend(this, params); }
+yy.BeginTransaction.prototype.toString = function() {
+	return K('BEGIN')+' '+K('TRANSACTION');
 }
 
-yy.Begin.prototype.execute = function (databaseid,params, cb) {
+yy.BeginTransaction.prototype.execute = function (databaseid,params, cb) {
 	var res = 1;
 	if(alasql.databases[alasql.useid].engineid) {
 		return alasql.engines[alasql.databases[alasql.useid].engineid].begin(databaseid, cb);
@@ -21,12 +21,12 @@ yy.Begin.prototype.execute = function (databaseid,params, cb) {
 	return res;
 };
 
-yy.Commit = function (params) { return yy.extend(this, params); }
-yy.Commit.prototype.toString = function() {
-	return K('COMMIT');
+yy.CommitTransaction = function (params) { return yy.extend(this, params); }
+yy.CommitTransaction.prototype.toString = function() {
+	return K('COMMIT')+' '+K('TRANSACTION');
 }
 
-yy.Commit.prototype.execute = function (databaseid,params, cb) {
+yy.CommitTransaction.prototype.execute = function (databaseid,params, cb) {
 	var res = 1;
 	if(alasql.databases[alasql.useid].engineid) {
 		return alasql.engines[alasql.databases[alasql.useid].engineid].commit(databaseid, cb);
@@ -37,12 +37,12 @@ yy.Commit.prototype.execute = function (databaseid,params, cb) {
 	return res;
 };
 
-yy.Rollback = function (params) { return yy.extend(this, params); }
-yy.Rollback.prototype.toString = function() {
-	return K('ROLLBACK');
+yy.RollbackTransaction = function (params) { return yy.extend(this, params); }
+yy.RollbackTransaction.prototype.toString = function() {
+	return K('ROLLBACK')+' '+K('TRANSACTION');
 }
 
-yy.Rollback.prototype.execute = function (databaseid,params,cb) {
+yy.RollbackTransaction.prototype.execute = function (databaseid,params,cb) {
 	var res = 1;
 	if(alasql.databases[alasql.useid].engineid) {
 		return alasql.engines[alasql.databases[alasql.useid].engineid].rollback(databaseid, cb);