+++ test/test203myfn.js
@@ -1,3 +1,0 @@
-alasql.fn.myfn = function(n){
-	return n*n;
-};